+++ c/src/pickle_table.c
@@ -3,7 +3,6 @@
 
 const PickleTable* PickleTable_new(const PickleRows* rows) {
     PickleTable* pickle_table = (PickleTable*)malloc(sizeof(PickleTable));
-    pickle_table->type = Argument_Table;
     pickle_table->rows = rows;
     return pickle_table;
 }