+++ javascript/src/GherkinLine.ts
@@ -1,5 +1,5 @@
-import countSymbols from './countSymbols'
-import type { IGherkinLine, Item } from './IToken'
+import countSymbols from './countSymbols.js'
+import type { IGherkinLine, Item } from './IToken.js'
 
 export default class GherkinLine implements IGherkinLine {
   public trimmedLineText: string
@@ -87,5 +87,3 @@ export default class GherkinLine implements IGherkinLine {
     return cells
   }
 }
-
-module.exports = GherkinLine