+++ javascript/test/GherkinInMarkdownTokenMatcherTest.ts
@@ -1,14 +1,14 @@
 import assert from 'node:assert'
-import type * as messages from '@cucumber/messages'
-import GherkinInMarkdownTokenMatcher from '../src/GherkinInMarkdownTokenMatcher'
-import GherkinLine from '../src/GherkinLine'
-import type { Item } from '../src/IToken'
-import type ITokenMatcher from '../src/ITokenMatcher'
-import { Token, TokenType } from '../src/Parser'
+import type { Location } from '@cucumber/messages'
+import GherkinInMarkdownTokenMatcher from '../src/GherkinInMarkdownTokenMatcher.js'
+import GherkinLine from '../src/GherkinLine.js'
+import type { Item } from '../src/IToken.js'
+import type ITokenMatcher from '../src/ITokenMatcher.js'
+import { Token, TokenType } from '../src/Parser.js'
 
 describe('GherkinInMarkdownTokenMatcher', () => {
   let tm: ITokenMatcher<TokenType>
-  let location: messages.Location
+  let location: Location
 
   beforeEach(() => {
     tm = new GherkinInMarkdownTokenMatcher('en')