+++ javascript/test/GherkinLineTest.ts
@@ -1,5 +1,5 @@
 import assert from 'node:assert'
-import GherkinLine from '../src/GherkinLine'
+import GherkinLine from '../src/GherkinLine.js'
 
 describe('GherkinLine', () => {
   describe('#getTableCells', () => {